Functional Project Completed

+++ app/src/main/java/com/example/expensetracker/model/Rates.kt
@@ -1,8 +1,8 @@
 package com.example.expensetracker.model
 
 data class Rates(
-    val TRY: Double,
-    val GBP: Double,
-    val EUR: Double,
-    val USD: Double
+    val TRY: Float,
+    val GBP: Float,
+    val EUR: Float,
+    val USD: Float
 )
